chore: refactor handlers of chatroom action BLoC
