+++ lib/packages/core/lib/src/blocs/chatroom_action/handler/hide_emoji_keyboard_handler.dart
@@ -1,0 +1,5 @@
+part of '../chatroom_action_bloc.dart';
+
+void _handleHideEmojiKeyboard(Emitter<LMChatroomActionState> emit) {
+  emit(LMChatHideEmojiKeyboardState());
+}
