+++ lib/packages/core/lib/src/blocs/chatroom_action/handler/mark_read_chatroom_handler.dart
@@ -1,0 +1,7 @@
+part of '../chatroom_action_bloc.dart';
+
+Future<void> _handleMarkReadChatroom(LMChatMarkReadChatroomEvent event) async {
+  // ignore: unused_local_variable
+  LMResponse response = await LMChatCore.client.markReadChatroom(
+      (MarkReadChatroomRequestBuilder()..chatroomId(event.chatroomId)).build());
+}
