+++ lib/packages/core/lib/src/blocs/chatroom_action/handler/show_emoji_keyboard_handler.dart
@@ -1,0 +1,8 @@
+part of '../chatroom_action_bloc.dart';
+
+void _handleShowEmojiKeyboard(
+    LMChatShowEmojiKeyboardEvent event, Emitter<LMChatroomActionState> emit) {
+  emit(LMChatShowEmojiKeyboardState(
+    conversationId: event.conversationId,
+  ));
+}
